primer/mcp-servers/mcp-server-github/requirements.txt
mcp>=1.0.0
python-dotenv>=1.0.0
PyGithub
